Gradle and Email Sender
Added javax.mail to the gradle so that emails can be sent and then created a class specifically made to send the emails. Most likely will add a credential method for password of the sending email, will need to make a sending email and will also need to add a recipient as that is currently a string.